requirements.txt
Flask==2.3.2
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
requests==2.31.0
rsa==4.9
Werkzeug==2.3.6
gunicorn